I think to 2 key conveniences here are:
`enum LogLevel {
case assert, error, warning, info, debug
}

enum FSStatus: String {
case exclude = "fs-exclude"
case mask = "fs-mask"
case unmask = "fs-unmask"
}`

...as well as the methods createFSEvent and log

Maybe I will remove them from this singleton since there is no good reason yet for it to be this way. The original reason is it would have a currentCustomEvent to hold the status of where the user is in the app, it would have a boolean value for userConsentGiven, etc...

For now I am abstracting the 4 items out of a class as they are something that can be copy/pasted into a given project.

LogLevel

FSStatus

fsCreateEvent

fsLog

fsModify